Revert bypass of reservation protection system in r2850.
Think this one through properly next time. This makes it appear that closing a reserved request is allowed. Also, it's making the reservation protection system useless, as this is EXACTLY what it was implemented to protect against, because people couldn't be trusted not to close a request that was reserved to someone else. 

You were one of those people who abused that trust, Prodego, and here you are adding a bypass to it.

If you have a problem with it, take it up with me.

FYI: there is a function to override the reservation system already. it's called mysql -e "UPDATE acc_pend SET pend_reserver = 0 WHERE pend_id = X;" . It's available to OQ and myself. It should practically never be needed. That's why it's intentionally not part of the tool interface.

git-svn-id: https://waca.svn.sourceforge.net/svnroot/waca/trunk/acc@2856 41b7b2fa-e25d-4c8f-a56d-f93ce3f2ce94