prevent multiple email addresses being used
This prevents the , character from being used in the local part of an email address. Prior to this, an email address could be constructed as such:

blah,test@stwalkerster.co.uk

This would result in the email being sent to multiple recipients:
blah@toolserver.org
test@stwalkerster.co.uk

This commit should prevent that by disallowing , in the local part.