Fix "Other requets from IP" bug on requests w/ XFF
Appears to have been introduced with the new XFF system.  This
particular bug affects any request with XFF headers that are trusted
further than one level down.

Another bug exists when all IPs in the XFF headers as well as the
origin are trusted proxies.  Not fixed in this commit.