Fixes bugs when all IP in a rq are trusted proxies
Fixes bugs when all IPs in a request and its XFF chain are listed as
trusted proxies; specifically, IP links not showing for origins that are
themselves trusted proxies, and the origin not being listed as the
originating IP when the origin is a trusted proxy itself.  This is an
edge case of sorts, but at least it will be handled gracefully.