email template PDO object
I renamed the database table so it fits in more nicely with the new PDO
objects (I'd rather not be making views on new tables!).

I've swapped LogClass onto PDO for EmailTemplate, to show you how easy
it is to retrieve a specific row, although I've not come up with a nice
way of adding arbitrary parameters to retrieve lots of objects yet
(though I do have an idea).