Add grants script
This script will automatically fix all of the relevant mysql grants on a way
we can be sure that the same grants are applied across different environments.

To run, grab yourself a privileged session on the database server, and run:

CALL SCHEMA_GRANTS_REFRESH("production", "'production'@'%'", "maintenance");

This should give a sane level of permissions which is currently needed by our
architecture. Future work elsewhere will allow us to pass in "web" to get a further
reduced set of permissions for the web UI, and a diferent set of permissions for
maintenance scripts.

As this is entirely in the database, it should be easy enough for us to include in the
standard deployment runbook. This is being added to the main repo so other
developers can use the same grants as in production, and so that it can be a developer
responsibility to think about what grants are actually needed, and not just deferring it
to the DBAs.