Fix encoding issue in redirection tool (#431)
An issue reported by DatGuy on IRC, where a username request for Foo&Bar
was getting passed to Google as /search?q=Foo&Bar instead of
/search?q=Foo%26Bar - this is caused by us using htmlentities instead of
urlencoding. We're now additionally using json_encode to create the
javascript string instead of manually wrapping the search term in quotes
to prevent escaping the string.