Swap WHOIS tools, add button to check IP with Stop Forum Spam
This commit swaps the WHOIS tools presented in the IP address
information section, such that the primary WHOIS tool is now the WMF
ToolForge WHOIS tool, and DomainTools is now the alternate tool because
CAPTCHAs are annoying.  Resolves #456.

In addition, this commit adds another IP check tool in addition to the
exiting Project Honeypot button, the Storm Forum Spam website.  Resolves
#457.