Perform sanity checks on XFF data of requests (#465)
Validate that the data we're being given in the X-Forwarded-For headers is
sane - specifically, only keep items in the chain that are actually valid IP
addresses, and silently discard anything that isn't.  Hopefully this will
resolve some deadlock errors we've been seeing on ACC recently, as the presence
of such an invalid XFF entry in a recent request was tripping up the geolocation
pre-cacher and thus causing all manner of undesired behavior.

Remember kids, don't trust user input!