Remove LIMIT clauses from write queries
Using LIMIT is unsafe in a replication environment due to the nature of
how MySQL retrieves rows to process the update might mean it retrieves a
different set of rows on the master and the slave, causing them to drift
out of sync.

This change removes this clause from all update and delete statements,
and by a happy coincidence all of these are already using the table's
primary key in the WHERE clause, so this is functionally a no-op, and
operationally hides a whole bunch of warnings in the log.