RequestData/alwaysSeeHash should be independent of alwaysSeePrivateData
This in theory allows us to remove private data visibility by default,
but still allow users with alwaysSeeHash to click the link to see
private data. This allows the possibility of better logging of who sees
what data.

Note that this makes alwaysSeeHash as powerful as alwaysSeePrivateData.
As these two grants are currently only granted together, this makes no
difference to current configuration.