Split off related requests visibility into separate right
This right adds the ability to see the existence of related requests
prior to reserving the request. When a request is reserved, the data is
visible whether or not the user has this right. When a user has the
alwaysSeePrivateData right, this also has no effect.

This is granted to tool users as a fix for a regression from the legacy
tool.

Fixes #568