Prevent non-admins from reopening purged requests
Using email as the test here, which is what the rest of the application
uses. This is locked behind a new user right; distinct from reopening
old non-purged requests.

I've specifically set this as a tool admin task, not a "request admin
tools" task, as checkusers should have no need to reopen a purged
request. Opening old requests I could see being useful for a CU, but
there's nothing that can realistically be done for a purged request.