Fix database migration scripts for development installations.
These changes have no effect on Production, partly because a) they've
already run, and b) they depend on the use of the default install seed
data.

Roughly speaking, these fixes are as follows:
* Update the default password to one that's actually able to log in
  after the changes to minimum password complexity rules
* Force the default admin user to be marked as identified by default
* Force the default admin user to use an automatic skin, which is the
  default at the moment
* Update the status of the default admin user to 'Active' - the change
  is normally done in the MigrateToRoles.php script, but this is not run
  when the database is installed from the base data.

I've also yanked the instructions from INSTALLING.md to sync git
submodules (we got rid of those), and removed the setting for
`$forceIdentification` as this does nothing now. I've also yanked the
instruction to manually set the user as Active - this is now handled in
the initial installation script.

Finally, I've also added *experimental* Docker support in the form of
a Dockerfile for the main application, and a docker-compose.yml file to
bring up both the application and the database. YMMV, it seems to work
for me, so it'll probably work for others, but you should know Docker
before playing with it.

Fixes #620