Allow usage of identity tickets for old OAuth consumer tokens
This will ease the transition from one consumer to another as existing
OAuth identities will still be considered "valid" by the application.

We permit the use of expired identity tickets in a *lot* of places in
the application, and this continues this tradition. If an expired ticket
can be used, it will be used instead of pulling the new token. This is
existing behaviour, though this change converts preferences to use the
cached ticket instead of forcing a new ticket to be retrieved (which
given the "refresh" button at the bottom of preferences, this is
expected behaviour anyway - the previous behaviour was IMHO a bug)

Effectively, this means that logged in users can continue to use the old
OAuth consumer for the duration of their login session. Other users will
continue to see the cached/expired identity ticket for the legacy
consumer. OAuth cache refresh jobs will not enforce a refresh of the
ticket for users using the old consumer - no further use of the old
consumer is permitted, so we must stick with the data we have.

That said, as soon as a user logs in (or presses "refresh" in prefs),
we will direct the user down the usual OAuth approval flow for the new
consumer, at which point they get an identity ticket on the new consumer
and all is well again. We just need to wait for everyone to log in after
the consumer change.

Failure to apply this change will result in statistics and preferences
being unavailable for all users still using the legacy consumer.

Related to #141 and #207