Fix cleanup of non-email-confirmed requests
The old version didn't take into account comments being left on requests
either by a requester or by tool users.

This meant that these requests would sit and not be deleted by the
script which runs every night to clean up non-confirmed requests. In
addition, because these requests were marked as "Open", these requests
wouldn't have private data cleared from them either.

Now we take the approach of deleting requester comments when there are
no other constraints which would prevent a request deletion, and when
there are constraints which prevent deletion, we instead mark the
request as closed, and "stale" to prevent an email confirmation from
succeeding. The number of requests this affects should be close to zero
though, because it requires a tool admin to explicitly override the
email-confirmation lockout and add a comment or affect a request in such
a way which adds a log entry (or a job queue entry).

We don't actually clear private data from "stale" requests immediately
either; private data will only be cleared from unconfirmed requests once
they are  MAX(email-confirmation timeout, private data purge timeout)
old, which I think is acceptable.