Fix automatic welcome prevention when welcoming is not possible
This fixes a theoretically rather obscure case (but prevalant currently)
where a user has welcoming enabled, but does not have the OAuth
configuration required for it. This means that welcoming is always
"skipped" from the user's perspective, as the "skip" checkbox is locked
on.

The form POST logic relied on checking the POSTed value to determine if
the welcome should be skipped due to the value of the checkbox. However,
since the checkbox was disabled, it's value is never sent to the server
as part of the form submission, thus the logic "saw" an unchecked
checkbox and proceeded with queueing a welcome task anyway.

This fixes it by re-running the check logic on POST as well as on GET.