Handle exceptions during pre-save validations better
We shouldn't kill the request due to a validation exception during the
pre-save validation step. Instead, we should allow a human to validate
the request manually - humans can deal with the occasional failure, and
this prevents a systematic failure of a validation method from killing
all inbound requests.

We need to log the exceptions, but I'll do that in a separate commit.

Fixes #656