Allow logging arbitrary exceptions
This builds on the exception log we already have for unhandled
exceptions, extending the exact same state logging but with the ability
to use this for arbitrarily-thrown exceptions. The intention here is not
to use this to "hide" errors when they actually matter nor to use this
extensively, but more to catch errors when it's more important to carry
on with the page request.

This change also enables logging for the validation requests from the
last commit.

Fixes #656