Cache role/identification lookups within the current request context
*This touches core security code*

Currently, user role and identification lookups are done every single
time we do an ACL check on the tool. For most pages, this happens about
twenty times per request, which is about forty database queries.

This change caches this information on a per-request basis, such that
lookups done *within the same request* come from the cache, causing the
database lookups to be done once per request, not once per ACL check.

The cache is split by user, for instances where we need to check another
account's ACLs (which is rare!). The only instance of this that comes to
mind is preference editing from within user management - we check the
target user's ACL to display whether or not a user is authorised to use
OAuth/bot creation. As such, the cache splitting can be fairly trivially
tested through the Edit button on User Management.

Verifying the cache is in use can only really be done with query logging
or full debugging.