Add badges to highlight concern areas (#746)
* Add badges to highlight concern areas

These badges are for the flagged comments and the job queue.  For flagged comments, any comments that have a flag are counted.  For the job queue, any failed and not acknowledged error is counted.  If we have either, it also shows a badge on the main navigation item to draw attention.

#735

* Update includes/Fragments/NavigationMenuAccessControl.php

Co-authored-by: Simon Walker <simon@stwalkerster.co.uk>

* Address code review comments

Co-authored-by: Simon Walker <simon@stwalkerster.co.uk>