Remove deprecated syntax from stylesheets
@extend for multiple selectors isn't stable, but the only uses are also
safe to just be converted into standard <input readonly> elements, so we
remove .uneditable-input entirely.

Additionally, there are issues with @import and media queries in modern
SCSS compilers. We avoid this by falling back to a plain CSS import.
This is slower as it causes the browser to load both stylesheets as
separate requests, rather than loading the combined set of stylesheets
in one larger request. To be frank, if the additional bandwidth is
really an issue for internal users, they can force a specific stylesheet
in their preferences and just load a single version.