Migrate from scssphp to dart-sass
This puts us closer to the standard SCSS implementation, and now we have
npm for popperjs we might as well use it for dart too.

Naturally, this means we need to use `npm run` instead of a PHP
maintenance script.

As a side effect, it's now possible to regenerate the stylesheets
without requiring a functional and up-to-data database, as the
application no longer needs to be initialised to do basic SCSS
operations.