Rework database installation script
Currently we pass the database password on the commandline. This is not
a Good Thing(tm), so this rewrites it to use a temporary configuration
file that is automatically deleted on exit.

I've also taken the opportunity to tidy up the commandline options to
something a little more user-friendly. The legacy approach still works,
but it's now undocumented with a preference towards moving to the new
syntax.