Implement global bans
Adds a "global" flag to the ban UI, though in practice this is just UI
sugar for a nullable domain column. If domain is null, then the ban
applies globally. If it's a domain ID, the ban applies in that domain
only.

Also adds a new ban type flag which is granted to tool roots only to
allow setting global bans.

Fixes #853