Implement privacy statement
This adds links to the WMCS TOS and our brand new {entry}.php/privacy
pages, which render the privacy statement in the correct context from a
markdown file. Links have been added to both public and private footers,
and to relevant data submission forms (request and register) for when
users start using the service.

I've also taken the opportunity to sassify the public stylesheet.