Add automatic comment flagging script
This hopefully should close the gap slightly between what humans forget
to flag and what we really should be flagging.

This is automated, and probably misses a lot. It's a copy-paste of the
query I've been using manually, so there's still a lot of things
commented out as they'll throw far too many false positives.

I fully expect this query to evolve over time.