Add configurable global deny role
This adds a configurable role definition which is applied to every ACL
check by everyone, logged in or not. This role definition is filtered to
only permit `deny` actions so additional access cannot be granted via
the configuration file.

The intent of this is to act as an emergency switch-off for any
functionality in the tool by denying everyone access to it.