Handle nulls in isUserIdentified
Currently, if a user's on-wiki username is not set, because they haven't
completed the required OAuth verification as part of the registration
process, for example, it's not possible to view the user's information/
statistics page because `isUserIdentified()` can't handle `null`s. This
addresses the issue; if passing in `null` or an empty string, it now
always returns `false`.